# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Apis",
        "description": "A simplest boilerplate for Laravel Apis",
        "contact": {
            "name": "Gilberto Junior",
            "email": "contact@codehappy.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/auth/refresh": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh the token.",
                "description": "Logged users are able to refresh the JWT Token.",
                "operationId": "refreshToken",
                "responses": {
                    "201": {
                        "description": "Created.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Token"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponse"
                                },
                                "example": {
                                    "message": "Unauthenticated."
                                }
                            }
                        }
                    }
                },
                "security": {
                    "bearerAuth": {}
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Token": {
                "description": "The response when the token was generated.",
                "required": [
                    "name",
                    "email"
                ],
                "properties": {
                    "message": {
                        "description": "The message from response.",
                        "type": "string"
                    },
                    "data": {
                        "description": "The token data.",
                        "properties": {
                            "access_token": {
                                "description": "The new JWT token.",
                                "type": "string"
                            },
                            "bearer_type": {
                                "description": "The type of header authorization.",
                                "type": "string"
                            },
                            "expires_in": {
                                "description": "The time of token expiration in seconds.",
                                "type": "integer",
                                "format": "int32"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "Token generated.",
                    "data": {
                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOlwvXC9hcGkudGVuYW50MS5oaXJlZDM2NS5kb2NrZXIuaW50ZXJuYWxcL2xvZ2luIiwiaWF0IjoxNjAzNDc2NTE5LCJleHAiOjE2MDM0ODAxMTksIm5iZiI6MTYwMzQ3NjUxOSwianRpIjoiTlNkUXI0M2gzQTFLMWJRSiIsInN1YiI6MSwicHJ2IjoiMWM5OGY5ZmNlMjZkM2JhNDg0NGNkYmQ5M2MzNjcwZTViNGJkYzlmOSJ9.fPbaYuFhOMdrJdMGyGO01_02BzeAUXMO0gHFyDqPcY4",
                        "token_type": "Bearer",
                        "expires_in": 3600
                    }
                }
            },
            "JsonResponse": {
                "description": "The structure of the api response.",
                "properties": {
                    "message": {
                        "description": "The message of the response.",
                        "type": "string"
                    },
                    "errors": {
                        "description": "The error messages of the response if exists.",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        {
            "bearerAuth": []
        }
    ]
}